refactor: make pytest testcases
