+++ docs/CHANGELOG.md
@@ -11,6 +11,11 @@
 - fix: support log debug level for load testing
 - fix: failed to load json/data content in api reference
 - fix: failed to convert postman collection containing multipart/form-data requests to pytest
+- fix: only get the first parameter in referenced testcase
+
+**python version**
+
+- refactor: make pytest testcases
 
 ## v4.1.4 (2022-06-17)
