+++ examples/data/a_b_c/T2_3_test.py
@@ -1,14 +1,12 @@
 # NOTE: Generated By HttpRunner v4.1.4
 # FROM: a-b.c/2 3.yml
-
+from httprunner import HttpRunner, Config, Step, RunRequest
+from httprunner import RunTestCase
 
 import sys
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).parent.parent))
-
-
-from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
 
 from a_b_c.T1_test import TestCaseT1 as T1
