+++ examples/httpbin/load_image_test.py
@@ -1,8 +1,6 @@
-# NOTE: Generated By HttpRunner v4.0.0
+# NOTE: Generated By HttpRunner v4.1.4
 # FROM: load_image.yml
-
-
-from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
+from httprunner import HttpRunner, Config, Step, RunRequest
 
 
 class TestCaseLoadImage(HttpRunner):
