+++ examples/httpbin/upload_test.py
@@ -1,8 +1,6 @@
-# NOTE: Generated By HttpRunner v4.0.0
+# NOTE: Generated By HttpRunner v4.1.4
 # FROM: upload.yml
-
-
-from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
+from httprunner import HttpRunner, Config, Step, RunRequest
 
 
 class TestCaseUpload(HttpRunner):
