+++ examples/postman_echo/request_methods/request_with_functions_test.py
@@ -1,8 +1,6 @@
 # NOTE: Generated By HttpRunner v4.1.4
 # FROM: request_methods/request_with_functions.yml
-
-
-from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
+from httprunner import HttpRunner, Config, Step, RunRequest
 
 
 class TestCaseRequestWithFunctions(HttpRunner):
