+++ examples/postman_echo/request_methods/request_with_parameters_test.py
@@ -1,14 +1,9 @@
 # NOTE: Generated By HttpRunner v4.1.4
 # FROM: request_methods/request_with_parameters.yml
-
-
 import pytest
 
-
+from httprunner import HttpRunner, Config, Step, RunRequest
 from httprunner import Parameters
-
-
-from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
 
 
 class TestCaseRequestWithParameters(HttpRunner):
