+++ examples/postman_echo/request_methods/request_with_testcase_reference_test.py
@@ -1,14 +1,12 @@
 # NOTE: Generated By HttpRunner v4.1.4
 # FROM: request_methods/request_with_testcase_reference.yml
-
+from httprunner import HttpRunner, Config, Step, RunRequest
+from httprunner import RunTestCase
 
 import sys
 from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).parent.parent))
-
-
-from httprunner import HttpRunner, Config, Step, RunRequest, RunTestCase
 
 from request_methods.request_with_functions_test import (
     TestCaseRequestWithFunctions as RequestWithFunctions,
